Update point.lua
Replaces a thread with a command for ped pointing. With pointing now being registered as a command you can utilize the RegisterKeyMapping native to bind to b which increases performance by not having a loop running until needed and giving the players the ability to rebind the key to any button they desire!