Removed handsup on bikes and motorcycles
You can't "handsup" on bikes and motorcycles since the animation is bugged when on those vehicle. Also, on bikes you can use "X" to do other things like use a melee weapon to hit others, etc. It was hard to use "X" on bikes since it's stopped steering and the animation was bugged :(