Handsup properly disables actions
For me, the current handsup did not disable aiming or shooting (or anything at all) when i was not in a vehicle. 
And when I was in a vehicle, it wasn't disabling aiming. So you could still shoot from a vehicle while having hands up.
Now everything works as intended and it does what it was claiming to do before.

PS: I'm not the greatest at this so if anything is wrong let me know. But for me it works as intended and does what it needs to do, unlike the current version of handsup.