bike/glitchy anim fix
this change retains the ability of a ped to enter a passenger seat without shuffling, but gets rid of the glitchy animation where they attempt to start shuffling and are stopped by the script. places the ped into the vehicle once when they first attempt to shuffle, then sets the flag to disable their attempt to shuffle.