Adding vehicle removal functions
I noticed that the ignore.lua as well as blacklisting and BlacklistedScenarios don't seem to remove some of the vehicles it's trying to do. For example if you stand in grove stree and /time 00 00 you will see a buzzard that spawns and patrols the grove street area. LISA still has a frogger and swift that lands and takes off on server start, crop dusintg and random AI flights still occur on the northern and east costs.

After awhile I started a standlone script that looks for various vehicles and removes them, it seems to work perfectly and as they are not on my blacklist I can still spawn them with /car and use them for scenarios in-city during roleplay.

I know this might be wrong, or there would be a better way to sort it, but figured it might be worth something to someone to add.