[FIX] Spam X to get out of vehicle, run, or aim
**Problem:**
When raising your hands, sometimes another animation plays and the player doesn't recognize the need to cancel the hands up action. This results in the player being unable to run, aim or drive. 

**Solution:**
This PR adds a listener to re-enable the disabled controls after having your hands raised. It's automatic by listening for the hands up animation to end. There is no ms added when not in use because the thread is only created when hands are up, and ends when the hands up animation ends or another animation plays.